+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.4.2](https://github.com/whale-labs/proto-converter/compare/v0.4.1...v0.4.2) (2020-11-26)
+
+## [0.4.1](https://github.com/whale-labs/proto-converter/compare/v0.4.0...v0.4.1) (2020-11-26)
+
+### Features
+
+- **plugins/nestjs/module:** add new module to the root module of the application
+- **graphql:** add default fields to empty response type
+- **graphql:** add default value to map field without comment
+- **types:** add default value to map field without comment
+
 ## [0.4.0](https://github.com/whale-labs/proto-converter/compare/v0.3.0...v0.4.0) (2020-11-24)
 
 ### Bug Fixes